# AULAS/Aula 06/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Controle de Filmes - SENAI Jandira
  description: |-
   API destinada a controle de filmes para o projeto "Bem Legal" do SENAI
  termsOfService: http://swagger.io/terms/
  contact:
    email: gustavovidalgva@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://
paths:
...
  /v1/senai/locadora/atividade:
    post:
      tags:
        - Atividade
      summary: Adiciona uma nova atividade
      description: Endpoint para enviar os dados da atividade a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova atividade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Atividade'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Atividade
      summary: Lista todos as atividades
      description: Endpoint para listar todos as atividades cadastradas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Atividade'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/atividade/{id}:
    get:
      tags:
        - Atividade
      summary: Busca uma atividade pelo id
      description: Endpoint para retornar uma atividade pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da atividade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Atividade'          
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Atividade
      summary: Atualiza uma atividade pelo id
      description: Endpoint para atualizar uma atividade pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da atividade
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma atividade existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Atividade'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Atividade
      summary: Exclui uma atividade pelo id
      description: Endpoint para excluir uma atividade pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da atividade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  
  # ENDPOINTS - ATOR
...
  /v1/senai/locadora/classificacao:
    post:
      tags:
        - Classificação
      summary: Adiciona uma nova classificação
      description: Endpoint para enviar os dados da classificação a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova classificação a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Classificação
      summary: Lista todos as classificações
      description: Endpoint para listar todos as classificações cadastradas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Classificacao'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/classificacao/{id}:
    get:
      tags:
        - Classificação
      summary: Busca uma classificação pelo id
      description: Endpoint para retornar uma classificação pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da classificação
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classificacao'          
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Classificação
      summary: Atualiza uma classificação pelo id
      description: Endpoint para atualizar uma classificação pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da classificação
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma classificação existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Classificação
      summary: Exclui uma classificação pelo id
      description: Endpoint para excluir uma classificação pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da classificação
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  
  # ENDPOINTS - DIRETOR
  /v1/senai/locadora/diretor:
    post:
      tags:
        - Diretor
      summary: Adiciona um novo diretor
      description: Endpoint para enviar os dados do diretor a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo diretor a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Diretor
      summary: Lista todos os diretores
      description: Endpoint para listar todos os diretores cadastrados
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiretorGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/diretor/{id}:
    get:
      tags:
        - Diretor
      summary: Busca um diretor pelo id
      description: Endpoint para retornar um diretor pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do diretor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiretorGet'          
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Diretor
      summary: Atualiza um diretor pelo id
      description: Endpoint para atualizar um diretor pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do diretor
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um diretor existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Diretor'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Diretor
      summary: Exclui um diretor pelo id
      description: Endpoint para excluir um diretor pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do diretor
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)

  # ENDPOINTS - FILME
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo filme
      description: Endpoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Filme
      summary: Lista todos os filmes
      description: Endpoint para listar todos os filmes cadastrados
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilmeGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um filme pelo id
      description: Endpoint para retornar um filme pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmeGet'          
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo id
      description: Endpoint para atualizar um filme pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um filme existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Filme
      summary: Exclui um filme pelo id
      description: Endpoint para excluir um filme pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)

  # ENDPOINTS - FOTO
  /v1/senai/locadora/foto:
    post:
      tags:
        - Foto
      summary: Adiciona uma nova foto
      description: Endpoint para enviar os dados da foto a serem cadastrados
      requestBody:
        description: JSON contendo os dados da nova foto a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Foto
      summary: Lista todas as fotos
      description: Endpoint para listar todas as fotos cadastradas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Foto'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/foto/{id}:
    get:
      tags:
        - Foto
      summary: Busca uma foto pelo id
      description: Endpoint para retornar uma foto pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Foto'          
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Foto
      summary: Atualiza uma foto pelo id
      description: Endpoint para atualizar uma foto pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da foto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma foto existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Foto
      summary: Exclui uma foto pelo id
      description: Endpoint para excluir uma foto pelo id
      parameters:
        - name: id
          in: path
          description: id para realizar a busca da foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)

  # ENDPOINTS - GÊNERO
  /v1/senai/locadora/genero:
    post:
      tags:
        - Gênero
      summary: Adiciona um novo gênero
      description: Endpoint para enviar os dados do gênero a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo gênero a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Gênero
      summary: Lista todos os gêneros
      description: Endpoint para listar todos os gêneros cadastrados
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genero'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/nacionalidade:
    post:
      tags:
        - Nacionalidade
      summary: Adiciona uma nova nacionalidade
      description: Endpoint para enviar os dados da nacionalidade a serem cadastradas
      requestBody:
        description: JSON contendo os dados da nova nacionalidade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Nacionalidade
      summary: Lista todas as nacionalidade
      description: Endpoint para listar todas as nacionalidade cadastrados
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Nacionalidade'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
components:
  schemas:
...
    Atividade:
      required:
        - id
        - area_atuacao
      type: object
      properties:
        id:
          type: integer
        area_atuacao:
          type: string
          description: O número máximo de caracteres é de 40
...
    Classificacao:
      required:
        - id
        - classificacao
      type: object
      properties:
        id:
          type: integer
        classificacao:
          type: string
          example: "+12"
          description: O número máximo de caracteres é de 5; Em caso de "Livre", se guarda a letra "L"
        descricao:
          type: string
          description: O número máximo de caracteres é ilimitado
        idade_minima:
          type: integer
    Diretor:
      required:
        - id
        - nome
        - data_nascimento
        - inicio_carreira
        - id_sexo
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          description: O número máximo de caracteres é de 100
        data_nascimento:
          type: string
          format: date
          example: "2000-01-01"
        inicio_carreira:
          type: string
          pattern: '^\d{4}$'
          example: "2020"
          description: O formato de ano é de no máximo 4 dígitos (AAAA)
        id_sexo:
          type: integer
...
    Foto:
      required:
        - id
        - foto
      type: object
      properties:
        id:
          type: integer
        foto:
          type: string
          example: "https://fotos.com/foto_do_ator"
          description: O número máximo de caracteres é de 255
...